Algorithms

BioSimulator.SSA — Type.

SSA

Gillespie's Direct Method (SSA). Simulates a system of coupled reactions and species by computing the time to the next reaction and searching on the CMF.

Internals

end_time: The termination time, supplied by a user.

t: The current simulation time.

source

BioSimulator.FRM — Type.

FRM

Gillespie's First Reaction Method. It is statistically equivalent to SSA, but more computationally expensive: it computes the time to the next reaction as the minimum waiting time relative to the next firing times of each reaction.

Internals

end_time: The termination time, supplied by a user.

t: The current simulation time.

source

BioSimulator.NRM — Type.

NRM

Gibson and Bruck's Next Reaction Method, statistically equivalent to SSA. It provides better computational efficiency on networks with loosely connected reactions.

Internals

end_time: The termination time, supplied by a user.

t: The current simulation time.

pq: A priority queue that sorts reaction according to the their next firing times.

source

BioSimulator.ODM — Type.

ODM

Optimized Direct Method. Similar to SSA, with the added benefit of sorting reactions according to their propensities over time. This improves the search on the CMF when selecting the next reaction to fire.

Internals

end_time: The termination time, supplied by a user.

t: The current simulation time.

source

BioSimulator.SAL — Type.

SAL

Step Anticipation-τ Leaping. An algorithm method that improves upon the accuracy of τ-leaping techniques by incorporating a first-order Taylor expansion.

Internals

end_time: The termination time, supplied by a user.

t: The current simulation time.

ϵ: A parameter controlling the size of leaps. Higher values allow for larger leaps, but may compromise the accuracy of results.

δ: A τ-leaping parameter used to switch between SSA and SAL. For example, if intensity < δ the algorithm carries out an ordinary SSA step to avoid negative species populations.

α: A parameter used to contract a τ-leap in the event of negative populations. Aggresive contraction (lower values) will bias sample paths.

dxdt: Rates of change for each species.

drdt: Rates of change for each reaction propensity.

events: The number of Poisson arrivals within a τ-leap interval.

source