Developers

The simulate Routine

The flexibility of the simulate routine allows one to easily implement additional algorithms:

function serial_simulate(...)

  init!(algorithm, Xt, r)

  for i in 1:max_trials
    copy!(Xt, X0)
    update_all_propensities!(a, r, Xt)
    reset!(algorithm, a)
    interval = 1

    while !done(algorithm)
      interval = update!(output, Xt, get_time(algorithm), interval, i)
      step!(algorithm, Xt, r)
    end

    interval = update!(output, Xt, get_time(algorithm), interval, i)
  end

  return output
end

Here is a breakdown of what happens:

init!(...): This performs any initial computations and data structure setup to avoid allocating memory within a loop.

for i in 1:max_trials: This loop iterates over independent simulations.

copy!(...): Reset the populations to their initial values.

reset!(...): Reset algorithm variables and data structures.

while !done(...): The main loop in the algorithm.

update!(...): Sample the state vector and record it as needed.

step!(...): Execute one step in the algorithm.

Thus, one simply needs to implement the init!, reset!, and step! functions for an algorithm.

The Algorithm Interface

BioSimulator.Algorithm — Type.

An abstract type that species a common interface for all algorithms implemented in BioSimulator.

All concrete Algorithm types should be immutable and avoid subtyping Algorithm directly; see ExactMethod and TauLeapMethod. The following are normally implemented for Algorithm:

get_time : Return the current simulation time.

end_time : Return the termination time specified by the user.

done : Has the simulation terminated? Defaults to time >= end_time.

init! : Initialize an Algorithm prior to simulation. Memory allocation should happen here because it occurs outside loops in the main simulation routine.

step! : Carry out a simulation step. This should (1) update simulation time, (2) update the species counts, (3) update propensities, and (4) update any relevant data structures.

source

The ExactMethod Interface

BioSimulator.ExactMethod — Type.

An ExactMethod is an abstract type for algorithms that are equivalent to Gillespie's method for stochastic simulation.

It subtypes the Algorithm type and provides:

reset! : This method defaults to resetting the interal simulation time for an ExactMethod after each trial.

source

The TauLeapMethod Interface

BioSimulator.TauLeapMethod — Type.

A TauLeapMethod is an abstract type for algorithms that utilize the τ-leaping strategy to accelerate simulation.

It subtypes the Algorithm type and provides:

events : Returns the number of reaction events.

reset! : This method defaults to resetting the internal simulation time for a TauLeapMethod after each trial.

source