Modeling

Modeling

Creating a Model

The Network type is the starting point of the modeling process in BioSimulator. It represents a collection related Species that interact through the rules defined by Reactions.

Interface

BioSimulator.Network — Type.

Network(id)

Construct an empty Network representing a system of interacting particles. A particle is represented by a Species, and an interaction is represented by a Reaction.

Add a Species or Reaction using <=:

m <= Species("X", 100) # Adds a Species
m <= Reaction("birth", 2.0, "X --> X + X")

Simulate a Network by calling simulate:

simulate(m, SSA, time=1000.0, epochs=1000, trials=100)

Visualize a Network as a Petri net:

visualize(m)

See also: Species, Reaction, simulate, visualize

Arguments

id: An string identifier for the Network.

source

BioSimulator.Species — Type.

Species(id, [value=0])

Define a Species with a name and initial population value.

source

BioSimulator.Reaction — Type.

Reaction(name, rate, formula)

Define a Reaction with a name, a stochastic rate constant, and a formula.

The formula should use Species from a Network. Multiple reactants are separated by a + symbol; similarly for products. Reactants and products are separated by a --> symbol. Refer to the examples below:

Examples

Reaction("birth", 1.0, "X –> X + X")

Reaction("death", 0.5, "X –> 0") # no products

Reaction("immigration", 0.5, "0 –> X") # no reactants

source

Running Simulations

BioSimulator.simulate — Function.

simulate{T}(model, [algorithm::Type{T}=SSA];
  time   :: AbstractFloat=1.0,
  epochs :: Integer=1,
  trials :: Integer=1,
  algvars...)

Simulate a Network using the given algorithm.

The simulation routine will run until the termination time and record the system state at evenly spaced epochs. This repeats for the given number of trials.

Arguments

model: The Network to simulate.

algorithm: The algorithm used to carry out simulation. One of SSA, FRM, NRM, ODM, or SAL.

Optional Arguments

time: The amount of time to simulate the model, in units of the model.

epochs: The number of times to sample the vector of counts.

trials: The number of independent realizations to generate.

kwargs: Additional keyword arguments specific to each algorithm.

source

Visualizing Results

Simulation output is stored as a 3D array within a PartialHistory type where

the i dimension spans the species in the Network,

the j dimension spans the epochs sampled, and

the k dimension spans the number of trials.

BioSimulator provides quick visualizations of sample and mean trajectories, distributions, and phase plots via the Plots package. Please see the Examples page for specific examples.

Manipulating and Exporting Data

Manipulating the simulation data directly can be cumbersome. BioSimulator uses the DataFrames package to extract results as an easy-to-use DataFrame. We note that Plots supports plotting DataFrame objects. One can write simulation results to disk directly, but we recommend using the interface provided by DataFrames.

GUI Generation

BioSimulator provides experimental automatic graphical user interface generation. One must specify the parameters to expose in the interface and provide default values where necessary.